+++ .gitignore
@@ -12,4 +12,4 @@
 /captures
 .externalNativeBuild
 .idea
-/inspetor/keystore.gradle
+/legiti/keystore.gradle
