+++ legiti/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.inspetor">
+          package="com.legiti">
     <application>
         <provider
                 android:authorities="${applicationId}.inspetorprovider"
-                android:name=".InspetorProvider"
+                android:name="com.com.legiti.InspetorProvider"
                 android:exported="false"
                 android:enabled="true" />
     </application>
