+++ legiti/src/main/java/com/legiti/helpers/EnumValue.kt
@@ -1,4 +1,4 @@
-package com.inspetor.helpers
+package com.legiti.helpers
 
 interface EnumValue {
     fun rawValue(): String
