+++ legiti/src/main/java/com/legiti/services/InspetorResourceService.kt
@@ -1,13 +1,13 @@
 //
 //  InspetorResourceService.kt
-//  inspetor-android-sdk
+//  com.legiti-android-sdk
 //
 //  Created by Matheus Sato on 12/4/19.
 //  Copyright © 2019 Inspetor. All rights reserved.
 //
-package com.inspetor.services
+package com.legiti.services
 
-import com.inspetor.helpers.*
+import com.legiti.helpers.*
 
 
 interface InspetorResourceService {
