+++ legiti/src/main/java/com/legiti/services/InspetorService.kt
@@ -1,11 +1,11 @@
 //
 //  InspetorService.kt
-//  inspetor-android-sdk
+//  com.legiti-android-sdk
 //
 //  Created by Matheus Sato on 12/4/19.
 //  Copyright © 2019 Inspetor. All rights reserved.
 //
-package com.inspetor.services
+package com.legiti.services
 
 import android.content.Context
