+++ legiti/src/main/res/values/strings.xml
