+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':inspetor'
+include ':legiti'
